+++ TangramKitDemo/IntegratedDemo/AllTestModel&View/AllTest1TableViewCell.swift
@@ -115,7 +115,7 @@ extension AllTest1TableViewCell
         self.rootLayout.tg_bottomPadding = 5
         self.rootLayout.tg_width.equal(.fill)
         self.rootLayout.tg_height.equal(.wrap)
-        self.rootLayout.tg_cacheEstimatedRect = true
+        self.rootLayout.tg_cacheEstimatedRect = true //这个属性只局限于在UITableViewCell中使用，用来优化tableviewcell的高度自适应的性能，其他地方请不要使用！！！
         self.contentView.addSubview(self.rootLayout)
         
         //如果您将布局视图作为子视图添加到UITableViewCell本身，并且同时设置了布局视图的宽度等于父布局的情况下，那么有可能最终展示的宽度会不正确。经过试验是因为对UITableViewCell本身的KVO监控所得到的新老尺寸的问题导致的这应该是iOS的一个BUG。所以这里建议最好是把布局视图添加到UITableViewCell的子视图contentView里面去。
@@ -174,7 +174,7 @@ extension AllTest1TableViewCell
         self.rootLayout.tg_bottomPadding = 5
         self.rootLayout.tg_width.equal(.fill)
         self.rootLayout.tg_height.equal(.wrap)
-        self.rootLayout.tg_cacheEstimatedRect = true
+        self.rootLayout.tg_cacheEstimatedRect = true //这个属性只局限于在UITableViewCell中使用，用来优化tableviewcell的高度自适应的性能，其他地方请不要使用！！！
         self.contentView.addSubview(self.rootLayout)
         
         /*
@@ -223,7 +223,7 @@ extension AllTest1TableViewCell
         self.rootLayout.tg_bottomPadding = 5
         self.rootLayout.tg_width.equal(.fill)
         self.rootLayout.tg_height.equal(.wrap)
-        self.rootLayout.tg_cacheEstimatedRect = true
+        self.rootLayout.tg_cacheEstimatedRect = true //这个属性只局限于在UITableViewCell中使用，用来优化tableviewcell的高度自适应的性能，其他地方请不要使用！！！
         self.contentView.addSubview(self.rootLayout)
  
         /*